fix: not entering insert mode when invoked from telescope results
I have a custom `my_find_file_in_project` function that I use to
navigate to the files in my current project.

This invokes telescope to display the files, and if a directory is
selected, it opens yazi to browse the directory.

The problem was that when yazi was opened for a directory, yazi would
open in normal mode, and would not accept any commands and just move the
neovim cursor instead.

This change fixes that issue for me.

You can see the current implementation here:

https://github.com/mikavilpas/dotfiles/blob/00c5719241d9ed09a48417645d92e2e3706ab534/.config/nvim/lua/plugins/telescope.lua?plain=1#L47