feat(lsp): apply changes to related files when a file is renamed
Previously the language server could easily get out of sync when a file
was renamed in yazi. The language server would e.g. have import
statements that referenced the old file name.

This feature only applies when a language server is running and the
server has support for renaming. What now happens is:

- neovim is open and a language server is running
- neovim opens yazi
- yazi renames (or moves) a file that's part of the current workspace
- yazi is closed
- neovim sends a request to the language server to rename the file
- the language server responds with a list of changes to other files
- neovim applies those changes to the other files but doesn't save
  changes automatically (at least not for me)

I tested this in a typescript project as well as two rust projects.
Everything worked nicely.