fix(lsp): only notifying lsp of renames for open buffers
This fixes an error in the following case:

- in a directory containing many files, have only one file open in
  neovim (with other files in the directory not open)
- open yazi
- rename the directory containing the files

At this point, the LSP server would receive a notification of the
rename only for the file that was open in neovim, and not for the
other files in the directory. This would cause the LSP server to
only apply changes related to the one open file.