fix(lsp): renaming only notified the lsp for the current file
In projects where there are many types of files, such as readme files
(markdown) and code files (let's say, typescript), there was an
inconvenience that the currently active lsp servers would not be
notified unless they were connected to the current buffer.

That means that if you were editing a markdown file and wanted to rename
a typescript file using yazi, the lsp server for typescript would not be
notified of the rename.