feat: <c-s> closes yazi and starts telescope.nvim's live_grep
The live_grep is constrained to the last directory yazi was in (that is,
the current yazi directory when entering this key combination).

This can be used to comfortably explore parts of the current project (on
the directory level) in an ad hoc fashion.

The search action can be configured in the user's configuration in case
they want to use some other method of searching (even another plugin
entirely), or customize it, or disable it completely.