fix(plugins): fix failure on repeated installation
When a plugin or flavor is first installed and later updated, the
link to the plugin or flavor's directory is already in place. This
caused an error when the plugin or flavor was installed again (updated).

Now, reinstalling checks if the symlink is already in place and does
nothing if it is.