perf: processing open buffers only processes normal buffers
Buffers are processed when renaming, deleting, and trashing files in
yazi. Previously all open buffers were processed, but now only normal
buffers are.

This actually doesn't have a big impact on performance. I just wanted to
have a `perf` commit in the history 😄. Perhaps in big projects it might
be noticeable, or if some future feature will require more performance.