fix: open buffers deleted in yazi were not closed
When a buffer was opened in neovim and being edited, and then yazi
deleted it, the buffer was not closed. I'm not sure what broke it.

The fix is to close the buffer asynchronously. I don't understand why
that fixes the issue, but it does.

`vim.await` is really cool. I found it here:
https://www.reddit.com/r/neovim/comments/1cv5q45/comment/l4s2ya0/?utm_source=share&utm_medium=web3x&utm_name=web3xcss&utm_term=1&utm_content=share_button