fix: exiting insert mode with "<esc><esc>"
For me, the following bug was happening:
- I start yazi and see yazi in the floating terminal
- I press "<esc><esc>" (esc twice)
- pressing j does not make yazi move the focus to the next file. Instead
  the terminal seems to be in normal mode and the cursor moves down.

This only happens once in the neovim session.

It was possible to work around this issue by pressing `i` to seemingly
enter insert mode again 😄

This commit works around the issue by always entering insert mode when
the mode changes.

This seems like a pretty stupid solution, so let's hope it causes no
problems 🤞🏻