feat: lazy.nvim users no longer need to specify dependencies (#144)
As of lazy.nvim 11, packages can now specify their own dependencies and
configuration. This means that users of lazy.nvim no longer need to
specify plenary.nvim as a dependency. If any dependencies are added in
the future, they will be automatically installed by lazy.nvim.

This is a backwards-compatible change.

<https://lazy.folke.io/news#11x>