fix(repro): update the repro instructions to update the dependencies
This plugin includes a `repro.lua` file that can be used to start neovim
with a minimal, known configuration. This is useful for reproducing issues.

If the reproduction script had been run some time ago, the dependencies
might be out of date. This commit updates the instructions with a
command that updates the dependencies before starting neovim. It should
ensure that the dependencies are up to date.