feat(plugin): allow specifying a subdirectory for plugins
Some plugins are not found at the root of the repository, but in a
subdirectory. This change allows specifying a subdirectory when
installing a plugin.

It corresponds to using `#` in the plugin URL when installing plugins
without yazi.nvim with the `ya pack` command.

```sh
ya pack -a redbeardymcgee/yazi-plugins#keyjump
```

Is the same as:

```lua
---@type LazySpec
{
  "redbeardymcgee/yazi-plugins",
  lazy = true,
  build = function(plugin)
    require("yazi.plugin").build_plugin(plugin, { sub_dir = "keyjump.yazi" })
  end,
},
```