feat: highlight the currently hovered file in yazi (opt-in) (#180)
For now, this change is opt-in (disabled by default).

This change makes it possible to highlight the currently hovered file with a
style of the user's own choosing.

To enable it, you need to do the following steps in your configuration:

- set `use_ya_for_events_reading = true`
- set these new settings in your configuration:

  ```lua
  ---@type LazySpec
  {
    "mikavilpas/yazi.nvim",
    -- ...other settings you might already have
    ---@type YaziConfig
    opts = {
      -- add these:
      use_ya_for_events_reading = true,
      highlight_groups = {
        hovered_buffer_background = { bg = "#363a4f" },
      },
      -- ...other settings you might already have
    }
  }
  ```

For now, the color needs to be configured manually. In the future we may have a
good default color. If you use catppuccin, you can find all the colors in the
palette here: <https://catppuccin.com/palette>. I used the `Surface 0` color
from my catppuccin macchiato palette.