fix: LSP renaming did not work in some cases
There were some edge cases that caused issues when renaming files with
LSP servers. Hopefully fix all of them by using the implementation
that's also shared by neotree and nvim-tree:
https://github.com/antosha417/nvim-lsp-file-operations

Thanks @chaozwn for suggesting the fix!

Closes <https://github.com/mikavilpas/yazi.nvim/issues/80>

This work was started in
https://github.com/mikavilpas/yazi.nvim/pull/190 where the idea is to
add a dependency on nvim-lsp-file-operations, but that PR has become
slightly stalled for now.

This PR is a temporary fix to this stalling by simply embedding the new
package inside yazi.nvim for now. When the issues in the PR have been
resolved, the embedding can be removed.