feat!: add default colors for hover highlighting (#194)
* feat!: add default colors for hover highlighting

Hover highlighting is now enabled by default
===
This works with the latest yazi when `use_ya_for_events_reading = true`
is specified. To learn more about enabling this, see
<https://github.com/mikavilpas/yazi.nvim/pull/152>

This will be the default behaviour when yazi 0.3 is released.

Hover highlighting colors no longer require manual configuration
===

This change adds default colors for the hover highlighting feature. The
logic supports both light and dark colorschemes:
- for dark colorschemes, the hovered buffer will be lighter in color
- for light colorschemes, it will be darker

The logic is borrowed from the `bufferline.nvim` plugin, which has a lot
of users and is well-maintained. This also adds a dependency to
bufferline.nvim, but as an optimization only its color utility functions
are used as late as possible, and even they are only used when the user
hasn't specified a custom highlight color.

BREAKING CHANGE: The optional `hovered_buffer_background` key in the
`YaziConfigHighlightGroups` has been renamed to `hovered_buffer`. This
change was made to better reflect the purpose of the key.

To migrate, run this 🙂

```vim
:%s/hovered_buffer_background/hovered_buffer/gc
```