fix: symlinked files cannot be highlighted when hovered
I experienced this issue when hovering over my yazi.nvim lazy.nvim spec.
It's symlinked to the actual file in the lazy.nvim repository.

These are the same file:

```sh
/Users/mikavilpas/dotfiles/.config/nvim/lua/plugins/my-file-manager.lua
/Users/mikavilpas/.config/nvim/lua/plugins/my-file-manager.lua
```

This commit fixes the issue by resolving the real path of the file
before
comparing it to the path of the buffer.

Closes https://github.com/mikavilpas/yazi.nvim/issues/195