feat(plugin): overwrite existing symlinks when installing plugins
This happened to me when this plugin was moved to a different hosting
provider:

- https://github.com/DreamMaoMao/keyjump.yazi
- https://gitee.com/DreamMaoMao/keyjump.yazi

After I changed my lazy.nvim spec to this, I got an error because the
symlink already existed:

```lua
{
  name = "keyjump.yazi",
  url = "https://gitee.com/DreamMaoMao/keyjump.yazi.git",
  lazy = true,
  build = function(plugin)
    require("yazi.plugin").build_plugin(plugin)
  end,
},
```

This change makes it so that the symlink is removed before creating a new
one. Realistically, this is what the user wants most of the time anyway.