fix: crash without newest yazi
There was a bug in opting in to the feature from 8114817 (feat: add
targeted communication with the yazi instance (opt-in) (#225)). Even if
you had not set `use_yazi_client_id_flag = true` in your config, the
newest yazi was still required.

This commit fixes that bug by removing the passing of the `yazi_id`
(roll forward style). Later changes can still utilize the feature and
build on top of this.