feat: add `yazi.toggle()` to continue from the last hovered file
This commit adds a new function `yazi.toggle()` that can be used to
pseudo resume the previous yazi session. An example mapping is added to
the README to show how to use it.

Note that this requires a very new yazi right now,
https://github.com/sxyazi/yazi/pull/1305 from 2024-07-18

You also need to set `use_ya_for_events_reading` and
`use_yazi_client_id_flag` to true in your config.