feat: pressing `tab` in yazi jumps to dir of next open split
    This feature is only supported for the latest yazi version. For now,
    you have to compile yazi from source to be able to use it. See the
    readme for instructions on how to do this.

If you have files open in split windows in Neovim, and have started
yazi, you can now press the `tab` key to make yazi jump to the directory
of the next open buffer. This is useful when you have multiple files
open in different directories and want to quickly jump between them.

Right now this has the following limitation: it can only jump to the
directory of the next open buffer, but not necessarily focus the buffer
in yazi. This is because yazi doesn't have a built-in way to jump to a
specific file, only a directory. This is something that might be
improved in the future.