feat!: allow customizing keymaps more clearly (#244)
* feat!: allow customizing keymaps more clearly

This adds a new configuration option `keymaps` that can be used to
customize the keybindings that are active when yazi is open and focused.
You might want to do this if you have a _yazi_ keymap that conflicts with
a yazi.nvim keymap, or if you just prefer different keybindings.

If you have your lua LSP server setup, you can now get nice
autocompletion for all of the available keymaps.

The current defaults have not changed, and are as follows:

```lua
keymaps = {
  open_file_in_vertical_split = '<c-v>',
  open_file_in_horizontal_split = '<c-x>',
  open_file_in_tab = '<c-t>',
  grep_in_directory = '<c-s>',
  cycle_open_buffers = '<tab>',
},
```

BREAKING CHANGE: If you, for some reason, relied on the fact that
`set_keymappings_function` removed all the built-in keymappings, you
will need to change your configuration. You can get the same behaviour
by setting `keymaps = false`. But realistically I think almost nobody
has done this, so it should be fine.

* fixup! feat!: allow customizing keymaps more clearly

* fixup! feat!: allow customizing keymaps more clearly