feat: add optional search and replace integration (grug-far.nvim)
This commit adds an integration with grug-far.nvim, a plugin that
provides a way to search and replace in the current working directory.
It was recently added in LazyVim as the default, and it's really easy to
use.

When you press the keybinding (default: `<c-g>`), yazi will close and
grug-far.nvim will open with the current directory prefilled in the
search field. This makes it easy to search and replace in that directory
only. Note that this only seems to work when the target directory is
under cwd, so it's not possible to search in a parent directory.

The integration is added as a keymap in the config. The keymap is set to
`<c-g>` by default, but it can be changed in the config.

To configure it, set one of these in your yazi.nvim config:

```lua
-- to disable using the keymap entirely
{
  keymaps = {
    replace_in_directory = false
  }
}

-- to change the keybinding to something else
{
  keymaps = {
    replace_in_directory = "<c-o>"
  }
}

-- advanced: to create a custom integration for yourself
{
  integrations = {
    replace_in_directory = function (directory)
      -- your custom implementation
    end
  },
}
```

<https://github.com/MagicDuck/grug-far.nvim>