perf: lazy load yazi.nvim modules by default
Previously yazi.nvim would load most of its modules up front when
required. This is not ideal as typically this delays Neovim's startup
time. Although the delay is measured in milliseconds, it might add up if
tens of other plugins need to be loaded too.

This commit changes the behavior so that yazi.nvim only loads the
modules when they are actually needed. The number of loaded modules has
changed from 17 to 7, but I did not measure the startup time difference.