fix: grug-far integration not being able to search outside of the cwd
There was a bug in the grug-far integration (the default search and
replace integration in yazi.nvim) where it was not able to search
outside of the current working directory. This was due to the fact that
the file path to search in was passed in `Files Filter` instead of in
`Flags`, which is the correct place to pass the file path.

https://github.com/MagicDuck/grug-far.nvim/issues/146