fix: renaming a file twice not updating the buffer name (#259)
When a file was renamed once, the buffer name was updated correctly. But
when the file was renamed a second time, the buffer name was not
updated.

I no longer remember why I had the `rename_or_close_buffer` function in
the first place, but this simpler implementation seems to work.

Anyway, this is quite a rare bug, so I don't expect a huge impact one
way or the other.