fix: close the floating terminal if it loses focus
The floating terminal that shows yazi running can lose focus when e.g.
clicked outside of. This commit adds an autocmd that closes the floating
terminal in these cases.

The reason for this is that it's inconvenient to navigate back to the
floating terminal to close it manually. Even if you did this, it was in
normal mode - not insert mode which is typically what you want to be in
when you're using yazi.