feat: can limit search and replace to selected files only
Previously the search and replace operation would be applied to all
files in the directory where the operation was started. This commit adds
a feature to limit the search and replace operation to only the files
that were visually selected (with `v` or `<space>` by default in yazi)
in the yazi view.

This allows fine tuning of the search and replace operation to only the
files that are relevant to the user.