feat: allow searching in selected files with telescope
> The telescope (search) integration in yazi.nvim is mapped to the
> `<c-s>` keymap by default.

Previously, telescope would be applied to all files in the directory
where the operation was started. This commit adds a feature to limit the
search to only the files that were visually selected (with `v` or
`<space>` by default in yazi) in the yazi view.

This allows fine tuning of the search operation to only the files that
are relevant to the user.