feat: `<c-y>` to copy relative path to selected file(s)
> This feature is available both when using `ya_for_events_reading and
> without it.
>
> If you get any errors, run `:healthcheck yazi` for more information.

When yazi is open, pressing `<c-y>` will copy the relative path of the
currently hovered file to the clipboard. If multiple files are
selected, the relative paths of all selected files will be copied, one
per line.

The relative path is calculated from the file/directory yazi was started
in, so it should do the right thing in most cases.

This feature requires GNU `realpath` on Linux and `grealpath` on macOS.
`grealpath` is available in the `coreutils` package on Homebrew. On most
Linux systems, `realpath` should be available by default.