test: allow different sizes for the terminal in the integration tests (#299)
Previously there was a limitation for the browser window. Only a single
size worked. Other sizes would cause neovim's text not to be shown at
all. Not sure why this happened, but it seems that if I pass the current
size of the terminal's textarea to the server, this issue is avoided.