perf: don't set up nvim-tree and neo-tree in lsp-file-operations
In yazi.nvim, we don't need these plugins. If the user had them in their
configuration, they were being set up which required some (minimal) work
to be done. This commit removes that work.

I could have removed the code entirely, but I decided to keep it in case
we want to depend on the official lsp-file-operations in the future.