feat: highlight buffers in the same directory (opt-out) (#351)
This commit adds a new configuration option,
`highlight_hovered_buffers_in_same_directory` (boolean), which will
highlight buffers that are in the same directory as the hovered buffer
using a less intensive color.

This provides an intuitive way to see that you are close to a window
that is open, but not quite there yet, which can help when you want to
copy paste files in yazi.

The setting is on by default.

If you don't like the default color, you can change it by setting
`highlight_groups.hovered_buffer_in_same_directory` in your config,
similarly to the already existing `highlight_groups.hovered_buffer`.