fix: starting telescope or grug-far when hovering a directory
Issue:
Given the following directory structure:

- root/
  - dir1/
    - file1
  - file2

When hovering `dir1/` and starting the search or replace operation, the
operation would only consider the files in `dir1/`. This is not what the
user would expect. The user would expect the operation to consider the
files in `root/`.

Solution:
When hovering a directory, we should consider the directory the hovered
directory is in. This is the same behavior as when hovering a file.