fix: not being able to delete a buffer when it is modified
Problem:

When a file is trashed (moved to the trash) or deleted (hard deleted) in
yazi, the buffer associated with the file is deleted. However, if the
buffer is modified, the buffer cannot be deleted. This is because the
buffer has unsaved changes.

Solution:

Force delete the buffer when it is being deleted, losing any unsaved
changes.

Originally I wanted to make sure no unsaved changes are lost, but I
noticed neo-tree does not care about unsaved changes when deleting a
file, so I decided to follow the same behavior.

Solves https://github.com/mikavilpas/yazi.nvim/issues/397