fix: closing split when opening directory
Problem:

When the user had two or more split windows open and opened a directory
in yazi, the current split window was closed. The buffer was not closed,
however, and the user could still see the buffer in the buffer list.

This is an annoying side effect that serves no real purpose.

Solution:

Don't close the current split window when opening a directory in yazi.