fix: opening directories in splits and tabs opening empty buffers
In yazi.nvim, it's possible to open files in new splits (horizontal and
vertical) and tabs.

Issue: when opening a directory in a split or tab, and when the setting
open_for_directories is set to true, the directory would be opened in a
new floating yazi window, and a useless split or tab would be opened
with the currently open buffer.

Solution: check if the chosen file is a directory before opening it in a
split or tab. If it's a directory, don't open it in a split or tab (do
nothing).