fix: `nvim .` sent extra `i` key to yazi
Issue
=

When opening yazi with `nvim .`, an extra `i` key was sent yazi. If the
user had set up a custom mapping for `i` in their yazi config, it would
be triggered when opening yazi with `nvim .`. In other cases this bug
would not get triggered.

Solution
=

Check if we're in the situation where the `i` key needs to be pressed
(to enter insert mode) and only then send the key.