feat: allow cycle_open_buffers (`tab`) to work for a single file
Issue
=====

Using `tab` to cycle through open buffers does not work when only one
buffer is open. I disabled it because I thought it was not useful in
that case, but actually it can still be useful in the following case:

- you have opened yazi and have a single buffer open in neovim (no
  splits)
- you have navigated to a different directory in the file explorer
- you have gotten lost
- you want to get back to the beginning with `tab`

Solution
========

Enable `tab` to cd to the single buffer's directory.

Solves https://github.com/mikavilpas/yazi.nvim/issues/447