feat: emit YaziRenamedOrMoved event when files are renamed or moved
Problem
=======

When files are renamed or moved in yazi, there is no way for other
plugins to know about this. If the user uses yazi.nvim along with other
plugins that need to react to files being renamed or moved, it cannot be
done. This means that other plugins that depend on yazi cannot react to
these changes.

Currently yazi.nvim can only react to LSP delete/rename events, but
plugins that operate outside of the LSP protocol are not supported.

Solution
========

Add a new Neovim autocmd/event called `YaziRenamedOrMoved` that is
emitted when files are renamed or moved. This event will contain a table
of the paths that were renamed or moved.

Previously we were already emitting a Neovim event for `YaziDDSHover` -
now we have two custom events.

Closes https://github.com/mikavilpas/yazi.nvim/issues/494