fix: YaziRenamedOrMoved events could not be published in practice
Issue
=====

When a file was renamed/moved, a YaziRenamedOrMoved event was supposed to be
published. However, the event was not actually published in practice due
to `E5560: nvim_exec_autocmds must not be called in a lua loop
callback`.

Solution
========

The solution was to schedule the event publishing in a `vim.schedule` block.
This way, the event publishing is deferred to the next event loop iteration,
which avoids the error.

For some reason, this error is not visible in unit tests, so I added an
integration test to make sure it doesn't break again.