fix: not being able to resolve the last_directory
Background
==========

The `last_directory` is the last directory that yazi was in before it
was closed. This is useful for various custom integrations that users
might want to do with yazi.

Issue
=====

The `last_directory` variable seemed to always be empty. It used to
work, but I think it broke when I added the following commit some time
ago:

> feat: expose the current working directory to keybindings (#471)
>
> https://github.com/mikavilpas/yazi.nvim/commit/445f4877d8b80f9d24ea1f9b890c878211878d63

The `last_directory` variable is being set currently, but using it still
depends on the old implementation that was in effect before the commit
above - it basically cannot be found because it's still expected to be
in the old place.

Solution
========

Use the new implementation that was introduced in the commit above.