perf(tests): use `cy.runExCommand` in all tests
Instead of typing into the terminal to run ex commands, use the
`cy.runExCommand` command. I added it to tui-sandbox today, and it uses
the neovim node api via a socket connection. This seems to speed up the
entire integration test suite by about 3-7 seconds (~47 s now on my
machine).

The speedup comes from the fact that we don't have to wait for the
characters to be typed into the terminal, and we don't have to wait for
the output to appear on the screen in order to assert on it.

It also makes the tests more reliable, as in some cases
`cy.typeIntoTerminal` cannot be used to assert that something does not
exist in the output (since we don't know when the output is done).