feat: using folke/snacks.nvim can preserve window layouts on deletes
Issue
=====

When yazi deletes a file, yazi.nvim also closes that buffer. However,
the default implementation in neovim is to also close the window/split
if one is currently open. This is distracting and annoying for users
(me).

Solution
========

If the https://github.com/folke/snacks.nvim plugin is installed,
yazi.nvim uses that to delete the buffer while preserving the window
layout.

If it's not installed, the old (annoying) behavior is used.

In the future, snacks.nvim might come a required dependency for
yazi.nvim.