feat: prevent conflicts with custom yazi config for `<enter>` (opt-in)
Issue
=====

For custom ways to open files, such as "open in new vertical split",
"open in new tab", etc., yazi.nvim uses the `<enter>` key to open the
file. This is not very robust because the user might have set a custom
keybinding for the `<enter>` key in their yazi config. In this case, the
custom keybinding would be triggered instead of the file being opened.

Solution
========

Instead of relying on the `<enter>` key, yazi.nvim can now use `ya emit
open` to make yazi open the file(s) that are currently selected. This
completely avoids the issue, but requires a recent version of yazi
(0.4.0 or later).

To opt into this behaviour, set the following in your config:

```lua
{
  -- example for lazy.nvim
  "mikavilpas/yazi.nvim",
  -- ... (other settings here) ...
  ---@type YaziConfig
  opts = {
    future_features = {
      ya_emit_open = true,
      -- 👆🏻 this is the new setting
    },
  },
}
```

This issue was found in
https://github.com/mikavilpas/yazi.nvim/issues/611 where
smart-enter.yazi was found to have this behaviour by default:
https://github.com/yazi-rs/plugins/tree/main/smart-enter.yazi#advanced