feat: can change neovim's cwd to the last opened directory (opt-in)
Some users have a workflow where they change the cwd a lot. This feature
adds an option to change neovim's cwd to the directory that was last
open when yazi was closed.

To opt in to this feature, set the `change_cwd_on_exit` option to
`"always"` in your yazi config. You can see an example in the README
file.

Right now only "always" is supported, which means the cwd is always
changed regardless of whether files were opened or not, but in the
future we might add more options.

Related issues and discussions:
- https://github.com/mikavilpas/yazi.nvim/issues/617
- https://github.com/mikavilpas/yazi.nvim/issues/83
- https://github.com/mikavilpas/yazi.nvim/discussions/95