feat: open yazi, hovering the visual selection file path
If you select a file path in visual mode and then press your yazi
keymapping, yazi will open and hover the file at that path.

How is this useful?

- sometimes you don't want to open a file in neovim because it will
  start LSP servers, other plugins, or other things that you don't need
  for your task
- you want to use yazi to move or rename the file, or copy its contents

NOTE: To use this feature, you will need to add a visual mode mapping
for yazi. You can find an example of this for lazy.nvim in the README
file.