refactor: remove support for old hacky way of opening files
The old way of opening files was to send a fake <enter> keypress to yazi
to simulate the user opening a file. This is pretty hacky, because
- the user might not be in a state in yazi where the fake keypress
  works at all. For example, they might be renaming a file. This is rare
  but too fragile to build any future features on top of.
- the user might have set a custom keybinding for the enter key. This
  would break the whole thing.

The new yazi feature `ya emit open` is used internally instead. The
feature works the same way but requires yazi `0.4.0`. This breaking
change was communicated in
https://github.com/mikavilpas/yazi.nvim/pull/664 but I kept it in for a
couple of weeks to smooth the transition a bit.

This was deprecated in https://github.com/sxyazi/yazi/pull/1979

Note that I copied most of this commit message from the previous commit
because I am lazy, so it might be wrong 🙂.