feat: allow processing all yazi events as soon as possible
Issue
=====

Some yazi events are processed only when yazi exits. These include the
following events:
- `rename` (a file was renamed)
- `bulk` (files were renamed using yazi's bulk rename feature)
- `move` (a file was moved)
- `delete` (you guessed it, a file was deleted)

The reason for this is that at the time I added support for handling
these event types, yazi was not yet able to provide event information
before it was completely closed.

Solution
========

Add a new configuration option, `process_events_live`, which defaults to
`false` for now. When this option is `true`, all event types are
processed as soon as possible.

It's recommended to add the https://github.com/folke/snacks.nvim plugin
so that yazi.nvim can use it to delete buffers. By default, Neovim
closes split windows when a buffer is deleted. I find this annoying, but
snacks solves this well.

If this feature works well, this might become the default behavior in
the future.