fix: `:Yazi toggle` not being able to hover a directory
Issue
=====

When using `:Yazi toggle` after previously having hovered a directory,
yazi is reopened hovering the contents of the directory instead of the
directory itself.

Solution
========

Wait for yazi to start, and use `ya reveal` to make yazi focus on the
directory. This takes 50-100ms of time which is very slightly janky, but
it's much better to have an 80% solution than a 0% solution.

Closes https://github.com/mikavilpas/yazi.nvim/issues/727