refactor(types): correctly allow passing a partial YaziConfig
Issue
=====

When using lua_ls types, the lua LSP would complain when using the
public yazi functions. They were meant to allow passing a partial
YaziConfig, but the type definition required everything to be present.

The behaviour is that any overrides to the default config are used when
calling the apis.

Solution
========

Allow passing a partial YaziConfig to the public functions.