refactor!: don't install snacks.nvim automatically
BREAKING CHANGE: to migrate, you may need to add snacks.nvim as a
dependency manually if your neovim does not already have it from another
source. You can copy it from the instructions in the readme file.

Issue
=====

Previously, snacks.nvim was installed automatically. This was done so
that users don't have to add it as a dependency manually. This caused
confusion for users:

- https://github.com/mikavilpas/yazi.nvim/issues/774
- https://github.com/mikavilpas/yazi.nvim/issues/783

It was also found that it's not possible to not install snacks.nvim if
it's added this way. This seems to bother some users.

Solution
========

As snacks.nvim is not a hard dependency yet, we can still require
installing it manually. This way, some users can choose to remove it if
they don't want it.

It might become a hard dependency in the future, though.

Closes https://github.com/mikavilpas/yazi.nvim/issues/783