refactor: simplify lsp renaming with embedded snacks implementation
Issue
=====

Previously, the lsp renaming was implemented using lsp-file-operations
from https://github.com/antosha417/nvim-lsp-file-operations/issues/35

Back then it made sense to use the lsp-file-operations plugin to handle
this because
- it was stable
- there was no good way to test and maintain a custom solution

However, right now lsp renaming is e2e tested in CI using
https://github.com/mikavilpas/tui-sandbox and it's much easier to
maintain. Also,
https://github.com/folke/snacks.nvim/blob/main/docs/rename.md is very
popular and has a concise implementation of the same feature.

Solution
========

Replace the lsp renaming implementation with a copy of the one from
snacks.nvim.

This reduces the amount of code that needs to be maintained.